Another structure refactor, separate the CLI instead of the core lib.
This means that `roboime-next`+`robime-next-cli` is preferred over
`roboime-next`+`roboime-next-core`, also because there may be more
interfaces in the future like `roboime-next-gui` or
`roboime-next-server`.

Having the main project be the CLI only seems to make the library more
annoying to use, since the CLI will never be depended on, it seems a
waste to use `roboime_next_core` everywhere while `roboime_next` is
already taken.

This only matters because it is not desirable to require the CLI
dependencies for the `roboime-next` library users.